+++ common/src/main/scala/explore/common/ConstraintSetObsQueries.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2021 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.observationtree
+package explore.common
 
 import cats.data.NonEmptyList
 import cats.effect.IO
@@ -89,9 +89,9 @@ object ConstraintSetObsQueries {
               cloudExtinction
               skyBackground
               waterVapor
-            }            
+            }
           }
-        }    
+        }
       }
     """
