+++ common/src/main/scala/explore/common/ObsQueries.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2021 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package explore.observationtree
+package explore.common
 
 import cats.data.NonEmptyList
 import cats.effect.IO
@@ -35,7 +35,7 @@ object ObsQueries {
               type: __typename
               ... on Target {
                 target_id: id
-              }              
+              }
               ... on Asterism {
                 asterism_id: id
               }
